+++ src/Port.php
@@ -12,7 +12,6 @@
  */
 namespace League\Uri\Components;
 
-use InvalidArgumentException;
 use League\Uri\Interfaces\Component as UriComponent;
 
 /**
@@ -41,7 +40,10 @@ class Port extends Component implements UriComponent
         }
 
         if (!is_int($data) || $data < 1 || $data > 65535) {
-            throw new InvalidArgumentException(sprintf('Expected port to be a int or null; received %s', gettype($data)));
+            throw new Exception(sprintf(
+                'Expected port to be a int or null; received %s',
+                gettype($data)
+            ));
         }
 
         return $data;